+++ src/components/Homepage/ReviewJobs/ReviewJobs.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, waitFor, fireEvent } from "@testing-library/react";
 import ReviewJobs from "./ReviewJobs";
 import { vi, describe, it, beforeEach, expect } from "vitest";
-import * as JobServices from "@services/JobServices.";
+import * as JobServices from "@services/JobServices";
 import * as useGetUserModule from "@hooks/useGetUser";
 import { useNavigate } from "react-router-dom";
